feat: make the markdown button copy the page content in markdown instead of mdx
