+++ package.json
@@ -43,6 +43,7 @@
     "short-uuid": "^4.2.2",
     "slugify": "^1.6.6",
     "sonner": "^1.4.3",
+    "turndown": "^7.2.1",
     "tw-merge": "^0.0.1-alpha.3",
     "typesense-instantsearch-adapter": "^2.8.0",
     "yarn": "^1.22.22"
@@ -55,6 +56,7 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@types/turndown": "^5.0.5",
     "autoprefixer": "^10.0.1",
     "dotenv": "^16.4.5",
     "eslint": "^8",
